Feature/choice and cancel buttons

+++ db/migrate/20130404214038_words_word_index_may_be_non_unique.rb
@@ -1,8 +1,13 @@
 class WordsWordIndexMayBeNonUnique < ActiveRecord::Migration
   def change
     change_table :words do |t|
-      t.remove_index name: 'index_entries_on_word'
-      t.index :word
+      if index_exists?(:words, :word, name: 'index_entries_on_word')
+        t.remove_index name: 'index_entries_on_word'
+      end
+
+      unless index_exists?(:words, :word, name: 'index_words_on_word')
+        t.index :word
+      end
     end
   end
 end
